refactor: Move reader generation out of weaver.cs
