+++ Assets/Mirror/Editor/Weaver/Readers.cs
@@ -1,7 +1,257 @@
 using System;
-public class Readers
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+
+
+namespace Mirror.Weaver
 {
-    public Readers()
+    public static class Readers
     {
+        const int MaxRecursionCount = 128;
+
+        public static MethodReference GetReadFunc(TypeReference variable, int recursionCount = 0)
+        {
+            if (recursionCount > MaxRecursionCount)
+            {
+                Weaver.Error("GetReadFunc recursion depth exceeded for " + variable.Name + ". Check for self-referencing member variables.");
+                return null;
+            }
+
+            if (Weaver.WeaveLists.readFuncs.ContainsKey(variable.FullName))
+            {
+                MethodReference foundFunc = Weaver.WeaveLists.readFuncs[variable.FullName];
+                if (foundFunc.ReturnType.IsArray == variable.IsArray)
+                {
+                    return foundFunc;
+                }
+            }
+
+            TypeDefinition td = variable.Resolve();
+            if (td == null)
+            {
+                Weaver.Error("GetReadFunc unsupported type " + variable.FullName);
+                return null;
+            }
+
+            if (variable.IsByReference)
+            {
+                // error??
+                Weaver.Error("GetReadFunc variable.IsByReference error.");
+                return null;
+            }
+
+            MethodDefinition newReaderFunc;
+
+            if (variable.IsArray)
+            {
+                TypeReference elementType = variable.GetElementType();
+                MethodReference elementReadFunc = GetReadFunc(elementType, recursionCount + 1);
+                if (elementReadFunc == null)
+                {
+                    return null;
+                }
+                newReaderFunc = GenerateArrayReadFunc(variable, elementReadFunc);
+            }
+            else
+            {
+                if (td.IsEnum)
+                {
+                    return Weaver.NetworkReaderReadInt32;
+                }
+
+                newReaderFunc = GenerateReadFunction(variable, recursionCount);
+            }
+
+            if (newReaderFunc == null)
+            {
+                Log.Error("GetReadFunc unable to generate function for:" + variable.FullName);
+                return null;
+            }
+            RegisterReadFunc(variable.FullName, newReaderFunc);
+            return newReaderFunc;
+        }
+
+        public static void RegisterReadFunc(string name, MethodDefinition newReaderFunc)
+        {
+            Weaver.WeaveLists.readFuncs[name] = newReaderFunc;
+            Weaver.WeaveLists.generatedReadFunctions.Add(newReaderFunc);
+
+            Weaver.ConfirmGeneratedCodeClass();
+            Weaver.WeaveLists.generateContainerClass.Methods.Add(newReaderFunc);
+        }
+
+        static MethodDefinition GenerateArrayReadFunc(TypeReference variable, MethodReference elementReadFunc)
+        {
+            if (!variable.IsArrayType())
+            {
+                Log.Error(variable.FullName + " is an unsupported array type. Jagged and multidimensional arrays are not supported");
+                return null;
+            }
+            string functionName = "_ReadArray" + variable.GetElementType().Name + "_";
+            if (variable.DeclaringType != null)
+            {
+                functionName += variable.DeclaringType.Name;
+            }
+            else
+            {
+                functionName += "None";
+            }
+
+            // create new reader for this type
+            MethodDefinition readerFunc = new MethodDefinition(functionName,
+                    MethodAttributes.Public |
+                    MethodAttributes.Static |
+                    MethodAttributes.HideBySig,
+                    variable);
+
+            readerFunc.Parameters.Add(new ParameterDefinition("reader", ParameterAttributes.None, Weaver.CurrentAssembly.MainModule.ImportReference(Weaver.NetworkReaderType)));
+
+            readerFunc.Body.Variables.Add(new VariableDefinition(Weaver.int32Type));
+            readerFunc.Body.Variables.Add(new VariableDefinition(variable));
+            readerFunc.Body.Variables.Add(new VariableDefinition(Weaver.int32Type));
+            readerFunc.Body.InitLocals = true;
+
+            ILProcessor worker = readerFunc.Body.GetILProcessor();
+
+            worker.Append(worker.Create(OpCodes.Ldarg_0));
+            worker.Append(worker.Create(OpCodes.Call, Weaver.NetworkReadUInt16));
+            worker.Append(worker.Create(OpCodes.Stloc_0));
+            worker.Append(worker.Create(OpCodes.Ldloc_0));
+
+            Instruction labelEmptyArray = worker.Create(OpCodes.Nop);
+            worker.Append(worker.Create(OpCodes.Brtrue, labelEmptyArray));
+
+            // return empty array
+            worker.Append(worker.Create(OpCodes.Ldc_I4_0));
+            worker.Append(worker.Create(OpCodes.Newarr, variable.GetElementType()));
+            worker.Append(worker.Create(OpCodes.Ret));
+
+            // create the actual array
+            worker.Append(labelEmptyArray);
+            worker.Append(worker.Create(OpCodes.Ldloc_0));
+            worker.Append(worker.Create(OpCodes.Newarr, variable.GetElementType()));
+            worker.Append(worker.Create(OpCodes.Stloc_1));
+            worker.Append(worker.Create(OpCodes.Ldc_I4_0));
+            worker.Append(worker.Create(OpCodes.Stloc_2));
+
+            // loop start
+            Instruction labelHead = worker.Create(OpCodes.Nop);
+            worker.Append(worker.Create(OpCodes.Br, labelHead));
+
+            // loop body
+            Instruction labelBody = worker.Create(OpCodes.Nop);
+            worker.Append(labelBody);
+            worker.Append(worker.Create(OpCodes.Ldloc_1));
+            worker.Append(worker.Create(OpCodes.Ldloc_2));
+            worker.Append(worker.Create(OpCodes.Ldelema, variable.GetElementType()));
+            worker.Append(worker.Create(OpCodes.Ldarg_0));
+            worker.Append(worker.Create(OpCodes.Call, elementReadFunc));
+            worker.Append(worker.Create(OpCodes.Stobj, variable.GetElementType()));
+            worker.Append(worker.Create(OpCodes.Ldloc_2));
+            worker.Append(worker.Create(OpCodes.Ldc_I4_1));
+            worker.Append(worker.Create(OpCodes.Add));
+            worker.Append(worker.Create(OpCodes.Stloc_2));
+
+            // loop while check
+            worker.Append(labelHead);
+            worker.Append(worker.Create(OpCodes.Ldloc_2));
+            worker.Append(worker.Create(OpCodes.Ldloc_0));
+            worker.Append(worker.Create(OpCodes.Blt, labelBody));
+
+            worker.Append(worker.Create(OpCodes.Ldloc_1));
+            worker.Append(worker.Create(OpCodes.Ret));
+            return readerFunc;
+        }
+
+        static MethodDefinition GenerateReadFunction(TypeReference variable, int recursionCount)
+        {
+
+            if (!Weaver.IsValidTypeToGenerate(variable.Resolve()))
+            {
+                return null;
+            }
+
+            string functionName = "_Read" + variable.Name + "_";
+            if (variable.DeclaringType != null)
+            {
+                functionName += variable.DeclaringType.Name;
+            }
+            else
+            {
+                functionName += "None";
+            }
+
+            // create new reader for this type
+            MethodDefinition readerFunc = new MethodDefinition(functionName,
+                    MethodAttributes.Public |
+                    MethodAttributes.Static |
+                    MethodAttributes.HideBySig,
+                    variable);
+
+            // create local for return value
+            readerFunc.Body.Variables.Add(new VariableDefinition(variable));
+            readerFunc.Body.InitLocals = true;
+
+            readerFunc.Parameters.Add(new ParameterDefinition("reader", ParameterAttributes.None, Weaver.CurrentAssembly.MainModule.ImportReference(Weaver.NetworkReaderType)));
+
+            ILProcessor worker = readerFunc.Body.GetILProcessor();
+
+            if (variable.IsValueType)
+            {
+                // structs are created with Initobj
+                worker.Append(worker.Create(OpCodes.Ldloca, 0));
+                worker.Append(worker.Create(OpCodes.Initobj, variable));
+            }
+            else
+            {
+                // classes are created with their constructor
+
+                MethodDefinition ctor = Resolvers.ResolveDefaultPublicCtor(variable);
+                if (ctor == null)
+                {
+                    Log.Error("The class " + variable.Name + " has no default constructor or it's private, aborting.");
+                    return null;
+                }
+
+                worker.Append(worker.Create(OpCodes.Newobj, ctor));
+                worker.Append(worker.Create(OpCodes.Stloc_0));
+            }
+
+            uint fields = 0;
+            foreach (FieldDefinition field in variable.Resolve().Fields)
+            {
+                if (field.IsStatic || field.IsPrivate)
+                    continue;
+
+                // mismatched ldloca/ldloc for struct/class combinations is invalid IL, which causes crash at runtime
+                OpCode opcode = variable.IsValueType ? OpCodes.Ldloca : OpCodes.Ldloc;
+                worker.Append(worker.Create(opcode, 0));
+
+                MethodReference readFunc = GetReadFunc(field.FieldType, recursionCount + 1);
+                if (readFunc != null)
+                {
+                    worker.Append(worker.Create(OpCodes.Ldarg_0));
+                    worker.Append(worker.Create(OpCodes.Call, readFunc));
+                }
+                else
+                {
+                    Weaver.Error("GetReadFunc for " + field.Name + " type " + field.FieldType + " no supported");
+                    return null;
+                }
+
+                worker.Append(worker.Create(OpCodes.Stfld, field));
+                fields++;
+            }
+            if (fields == 0)
+            {
+                Log.Warning("The class / struct " + variable.Name + " has no public or non-static fields to serialize");
+            }
+
+            worker.Append(worker.Create(OpCodes.Ldloc_0));
+            worker.Append(worker.Create(OpCodes.Ret));
+            return readerFunc;
+        }
+
     }
+
 }
