+++ Assets/Mirror/Editor/Weaver/Weaver.cs
@@ -49,12 +49,6 @@ namespace Mirror.Weaver
         // private properties
         static bool DebugLogEnabled = true;
 
-        // this is used to prevent stack overflows when generating serialization code when there are self-referencing types.
-        // All the utility classes use GetWriteFunc() to generate serialization code, so the recursion check is implemented there instead of in each utility class.
-        // A NetworkBehaviour with the max SyncVars (64) can legitimately increment this value to 65 - so max must be higher than that
-        const int MaxRecursionCount = 128;
-        static int RecursionCount;
-
         // Network types
         public static TypeReference NetworkBehaviourType;
         public static TypeReference NetworkBehaviourType2;
@@ -179,12 +173,6 @@ namespace Mirror.Weaver
         public static MethodReference sendTargetRpcInternal;
         public static MethodReference sendEventInternal;
 
-        [Obsolete("Use Writer.ResetRocursionCount")]
-        public static void ResetRecursionCount()
-        {
-            RecursionCount = 0;
-        }
-
         public static void DLog(TypeDefinition td, string fmt, params object[] args)
         {
             if (!DebugLogEnabled)
@@ -211,250 +199,6 @@ namespace Mirror.Weaver
         public static void SetNumSyncVars(string className, int num)
         {
             WeaveLists.numSyncVars[className] = num;
-        }
-
-
-
-        public static MethodReference GetReadFunc(TypeReference variable)
-        {
-            if (WeaveLists.readFuncs.ContainsKey(variable.FullName))
-            {
-                MethodReference foundFunc = WeaveLists.readFuncs[variable.FullName];
-                if (foundFunc.ReturnType.IsArray == variable.IsArray)
-                {
-                    return foundFunc;
-                }
-            }
-
-            TypeDefinition td = variable.Resolve();
-            if (td == null)
-            {
-                Error("GetReadFunc unsupported type " + variable.FullName);
-                return null;
-            }
-
-            if (variable.IsByReference)
-            {
-                // error??
-                Error("GetReadFunc variable.IsByReference error.");
-                return null;
-            }
-
-            MethodDefinition newReaderFunc;
-
-            if (variable.IsArray)
-            {
-                TypeReference elementType = variable.GetElementType();
-                MethodReference elementReadFunc = GetReadFunc(elementType);
-                if (elementReadFunc == null)
-                {
-                    return null;
-                }
-                newReaderFunc = GenerateArrayReadFunc(variable, elementReadFunc);
-            }
-            else
-            {
-                if (td.IsEnum)
-                {
-                    return NetworkReaderReadInt32;
-                }
-
-                newReaderFunc = GenerateReadFunction(variable);
-            }
-
-            if (newReaderFunc == null)
-            {
-                Log.Error("GetReadFunc unable to generate function for:" + variable.FullName);
-                return null;
-            }
-            RegisterReadFunc(variable.FullName, newReaderFunc);
-            return newReaderFunc;
-        }
-
-        public static void RegisterReadFunc(string name, MethodDefinition newReaderFunc)
-        {
-            WeaveLists.readFuncs[name] = newReaderFunc;
-            WeaveLists.generatedReadFunctions.Add(newReaderFunc);
-
-            ConfirmGeneratedCodeClass();
-            WeaveLists.generateContainerClass.Methods.Add(newReaderFunc);
-        }
-
-        static MethodDefinition GenerateArrayReadFunc(TypeReference variable, MethodReference elementReadFunc)
-        {
-            if (!variable.IsArrayType())
-            {
-                Log.Error(variable.FullName + " is an unsupported array type. Jagged and multidimensional arrays are not supported");
-                return null;
-            }
-            string functionName = "_ReadArray" + variable.GetElementType().Name + "_";
-            if (variable.DeclaringType != null)
-            {
-                functionName += variable.DeclaringType.Name;
-            }
-            else
-            {
-                functionName += "None";
-            }
-
-            // create new reader for this type
-            MethodDefinition readerFunc = new MethodDefinition(functionName,
-                    MethodAttributes.Public |
-                    MethodAttributes.Static |
-                    MethodAttributes.HideBySig,
-                    variable);
-
-            readerFunc.Parameters.Add(new ParameterDefinition("reader", ParameterAttributes.None, CurrentAssembly.MainModule.ImportReference(NetworkReaderType)));
-
-            readerFunc.Body.Variables.Add(new VariableDefinition(int32Type));
-            readerFunc.Body.Variables.Add(new VariableDefinition(variable));
-            readerFunc.Body.Variables.Add(new VariableDefinition(int32Type));
-            readerFunc.Body.InitLocals = true;
-
-            ILProcessor worker = readerFunc.Body.GetILProcessor();
-
-            worker.Append(worker.Create(OpCodes.Ldarg_0));
-            worker.Append(worker.Create(OpCodes.Call, NetworkReadUInt16));
-            worker.Append(worker.Create(OpCodes.Stloc_0));
-            worker.Append(worker.Create(OpCodes.Ldloc_0));
-
-            Instruction labelEmptyArray = worker.Create(OpCodes.Nop);
-            worker.Append(worker.Create(OpCodes.Brtrue, labelEmptyArray));
-
-            // return empty array
-            worker.Append(worker.Create(OpCodes.Ldc_I4_0));
-            worker.Append(worker.Create(OpCodes.Newarr, variable.GetElementType()));
-            worker.Append(worker.Create(OpCodes.Ret));
-
-            // create the actual array
-            worker.Append(labelEmptyArray);
-            worker.Append(worker.Create(OpCodes.Ldloc_0));
-            worker.Append(worker.Create(OpCodes.Newarr, variable.GetElementType()));
-            worker.Append(worker.Create(OpCodes.Stloc_1));
-            worker.Append(worker.Create(OpCodes.Ldc_I4_0));
-            worker.Append(worker.Create(OpCodes.Stloc_2));
-
-            // loop start
-            Instruction labelHead = worker.Create(OpCodes.Nop);
-            worker.Append(worker.Create(OpCodes.Br, labelHead));
-
-            // loop body
-            Instruction labelBody = worker.Create(OpCodes.Nop);
-            worker.Append(labelBody);
-            worker.Append(worker.Create(OpCodes.Ldloc_1));
-            worker.Append(worker.Create(OpCodes.Ldloc_2));
-            worker.Append(worker.Create(OpCodes.Ldelema, variable.GetElementType()));
-            worker.Append(worker.Create(OpCodes.Ldarg_0));
-            worker.Append(worker.Create(OpCodes.Call, elementReadFunc));
-            worker.Append(worker.Create(OpCodes.Stobj, variable.GetElementType()));
-            worker.Append(worker.Create(OpCodes.Ldloc_2));
-            worker.Append(worker.Create(OpCodes.Ldc_I4_1));
-            worker.Append(worker.Create(OpCodes.Add));
-            worker.Append(worker.Create(OpCodes.Stloc_2));
-
-            // loop while check
-            worker.Append(labelHead);
-            worker.Append(worker.Create(OpCodes.Ldloc_2));
-            worker.Append(worker.Create(OpCodes.Ldloc_0));
-            worker.Append(worker.Create(OpCodes.Blt, labelBody));
-
-            worker.Append(worker.Create(OpCodes.Ldloc_1));
-            worker.Append(worker.Create(OpCodes.Ret));
-            return readerFunc;
-        }
-
-        static MethodDefinition GenerateReadFunction(TypeReference variable)
-        {
-            if (RecursionCount++ > MaxRecursionCount)
-            {
-                Weaver.Error("GetReadFunc recursion depth exceeded for " + variable.Name + ". Check for self-referencing member variables.");
-                return null;
-            }
-
-            if (!IsValidTypeToGenerate(variable.Resolve()))
-            {
-                return null;
-            }
-
-            string functionName = "_Read" + variable.Name + "_";
-            if (variable.DeclaringType != null)
-            {
-                functionName += variable.DeclaringType.Name;
-            }
-            else
-            {
-                functionName += "None";
-            }
-
-            // create new reader for this type
-            MethodDefinition readerFunc = new MethodDefinition(functionName,
-                    MethodAttributes.Public |
-                    MethodAttributes.Static |
-                    MethodAttributes.HideBySig,
-                    variable);
-
-            // create local for return value
-            readerFunc.Body.Variables.Add(new VariableDefinition(variable));
-            readerFunc.Body.InitLocals = true;
-
-            readerFunc.Parameters.Add(new ParameterDefinition("reader", ParameterAttributes.None, CurrentAssembly.MainModule.ImportReference(NetworkReaderType)));
-
-            ILProcessor worker = readerFunc.Body.GetILProcessor();
-
-            if (variable.IsValueType)
-            {
-                // structs are created with Initobj
-                worker.Append(worker.Create(OpCodes.Ldloca, 0));
-                worker.Append(worker.Create(OpCodes.Initobj, variable));
-            }
-            else
-            {
-                // classes are created with their constructor
-
-                MethodDefinition ctor = Resolvers.ResolveDefaultPublicCtor(variable);
-                if (ctor == null)
-                {
-                    Log.Error("The class " + variable.Name + " has no default constructor or it's private, aborting.");
-                    return null;
-                }
-
-                worker.Append(worker.Create(OpCodes.Newobj, ctor));
-                worker.Append(worker.Create(OpCodes.Stloc_0));
-            }
-
-            uint fields = 0;
-            foreach (FieldDefinition field in variable.Resolve().Fields)
-            {
-                if (field.IsStatic || field.IsPrivate)
-                    continue;
-
-                // mismatched ldloca/ldloc for struct/class combinations is invalid IL, which causes crash at runtime
-                OpCode opcode = variable.IsValueType ? OpCodes.Ldloca : OpCodes.Ldloc;
-                worker.Append(worker.Create(opcode, 0));
-
-                MethodReference readFunc = GetReadFunc(field.FieldType);
-                if (readFunc != null)
-                {
-                    worker.Append(worker.Create(OpCodes.Ldarg_0));
-                    worker.Append(worker.Create(OpCodes.Call, readFunc));
-                }
-                else
-                {
-                    Weaver.Error("GetReadFunc for " + field.Name + " type " + field.FieldType + " no supported");
-                    return null;
-                }
-
-                worker.Append(worker.Create(OpCodes.Stfld, field));
-                fields++;
-            }
-            if (fields == 0)
-            {
-                Log.Warning("The class / struct " + variable.Name + " has no public or non-static fields to serialize");
-            }
-
-            worker.Append(worker.Create(OpCodes.Ldloc_0));
-            worker.Append(worker.Create(OpCodes.Ret));
-            return readerFunc;
         }
 
         static internal void ConfirmGeneratedCodeClass()
